fix: stop collecting when failing to gather events
